Reorganization and CUDA fleshing.
