+++ docs/source/templates/custom-class-template.rst
@@ -6,7 +6,7 @@
    :members:
    :show-inheritance:
    :inherited-members:
-   :special-members: __call__, __add__, __mul__
+   :special-members: __call__, __add__, __mul__, __init__
 
    {% block methods %}
    {% if methods %}
